Convert InputChanged to short, make Math.Abs safe

+++ UCR.Core/Managers/BindingManager.cs
@@ -110,7 +110,7 @@ namespace HidWizards.UCR.Core.Managers
             };
         }
 
-        private void InputChanged(ProviderDescriptor providerDescriptor, DeviceDescriptor deviceDescriptor, BindingReport bindingReport, int value)
+        private void InputChanged(ProviderDescriptor providerDescriptor, DeviceDescriptor deviceDescriptor, BindingReport bindingReport, short value)
         {
             if (!DeviceBinding.MapCategory(bindingReport.Category).Equals(_deviceBinding.DeviceBindingCategory)) return;
             if (!IsInputValid(bindingReport.Category, value)) return;
@@ -121,9 +121,7 @@ namespace HidWizards.UCR.Core.Managers
             EndBindMode();
         }
 
-        // ToDo: This should take a short, not an int
-        // As it is currently an int holding a short, the Math.Abs statements are safe
-        private bool IsInputValid(BindingCategory bindingCategory, int value)
+        private bool IsInputValid(BindingCategory bindingCategory, short value)
         {
             switch (DeviceBinding.MapCategory(bindingCategory))
             {
@@ -133,8 +131,8 @@ namespace HidWizards.UCR.Core.Managers
                 case DeviceBindingCategory.Momentary:
                     return value != 0;
                 case DeviceBindingCategory.Range:
-                    return Constants.AxisMaxValue * 0.4 < Math.Abs(value)
-                        && Constants.AxisMaxValue * 0.6 > Math.Abs(value);
+                    return Constants.AxisMaxValue * 0.4 < Math.Abs((int)value)
+                        && Constants.AxisMaxValue * 0.6 > Math.Abs((int)value);
                 default:
                     return false;
             }
